requirements.txt
pandas
Pillow
torch
transformers
peft
accelerate